+++ src/main/java/org/support/project/web/dao/gen/GenGroupsDao.java
@@ -300,20 +300,4 @@ public class GenGroupsDao extends AbstractDao {
 		activation(entity.getGroupId());
 
 	}
-	/**
-	 * データをtruncateする
-	 */
-	@Aspect(advice=org.support.project.ormapping.transaction.Transaction.class)
-	public void truncate() {
-		String sql = SQLManager.getInstance().getSql("/org/support/project/web/dao/sql/GroupsDao/GroupsDao_truncate.sql");
-		executeUpdate(sql);
-	}
-	/**
-	 * sequenceをリセットする
-	 */
-	@Aspect(advice=org.support.project.ormapping.transaction.Transaction.class)
-	public void resetSequence() {
-		String sql = SQLManager.getInstance().getSql("/org/support/project/web/dao/sql/GroupsDao/GroupsDao_alter_sequence.sql");
-		executeUpdate(sql);
-	}
 }
