+++ src/main/java/org/support/project/web/dao/gen/GenSystemsDao.java
@@ -282,12 +282,4 @@ public class GenSystemsDao extends AbstractDao {
 		activation(entity.getSystemName());
 
 	}
-	/**
-	 * データをtruncateする
-	 */
-	@Aspect(advice=org.support.project.ormapping.transaction.Transaction.class)
-	public void truncate() {
-		String sql = SQLManager.getInstance().getSql("/org/support/project/web/dao/sql/SystemsDao/SystemsDao_truncate.sql");
-		executeUpdate(sql);
-	}
 }
